backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
python-dotenv
pydantic>=2.0
praw
twikit